Fix useFreesound hook to avoid extra re-render
Making the effect containing the access token call depend on search params is fickle. When the the token state is updated, a re-render is triggered which also seems to count for an update to the search params which again triggers the effect to redundantly fetch the token again since the auth code is still set.

However, it seems that Freesound has an issue with requesting an access token with the same auth code twice, causing things to fail and the token state to clear.

The workaround I have right now is to use a cleanup function to clear the auth code from the search params after fetching the token. This way when the effect is triggered again, it won't do another fetch.

The true solution here is to place this server side and use cookies to store the token... not fully fleshed out.

The downside right now is that after logging in, there's a brief period of time before the UI updates to indicate that the token is present. I could look into adding a loading UI state however.