Register fastifyJwt; add AES encryption functions
The plan is to encrypt the access and refresh tokens and supply them in a JWT. This allows Raudi to have stateless auth while never exposing Freesound tokens.

Raudi will only accept a full JWT, and is protected from XSS since there is no room whatsoever for code injection. CSRF attacks are also protected against since the JWT will be stored in a `SameSite` cookie.