Cleanup auth flow; add JWT to requests to sounds
`auth.ts` exposes a `useFreesound` hook which relevant routes can use to enforce that the user is logged into Freesound and get access to their Freesound access token.

I still need to handle lots of edge cases like if decryption fails or the access token expires. Otherwise, this is a working proof of concept for the auth flow.