Piece auth back together
Adding in the `me` endpoint revealed a lot of issues with the current auth flow. This brings it to a semi-working state (not all edge cases are covered yet). Much of the issues here seem to stem from quirks with using a `purify-ts` workflow with `fastify`, for instance calling `reply` functions within a `caseOf` will be overridden with a 200 unless the `Either`/`Maybe` being used is returned by the handler.