Cleanup `useJwt` and handle error cases
This commit introduces a helper for chaining computations that return a nullable value and may throw an exception. This be-all-end-all helper can cover all cases and coerce them into a `Maybe` which is simply joined with the previous computation.

I have future plans to create a version of this for `Either` when the need arises.