POC login check
Client pages requiring login will verify login by pinging the `me` endpoint on render, redirecting to the login page on failure to authenticate.

Conversely, visiting the login page will redirect you to the main page if you are already logged in.

Currently both pages ping independently, but this is wasteful and will result in unnecessary pings. To avoid this, user login state should be hoisted to the app level, initialized as `false` and set to `true` upon successful login confirmation. If any request in the future returns unauthorized, then this state should be set back to false and the user should be redirected to the login.

This logic can be factored into a hook so that all guarded pages can reuse this logic.