Add login context; create `useLogin` hook
The main root component now has login state which will be passed down to child components via a login context.

The login context itself is a state machine with three states:
1. Unknown
2. Logged In
3. Logged Out

When the user first visits the sight, their login state is `Unknown`. The client must make a request to the server specifically to see whether the user is logged in. At this point, the state transitions to either `Logged In` or `Logged Out` depending on the response.

Then, the user may switch between these two states depending on of they get any unauthorized requests back or re-login.

This logic (or at least most of it, there is no abstraction yet for setting the user to be logged out when getting back unauthorized responses from the server) lives within a `useLogin` hook which hides the details of this state machine, allowing components to declaratively both require login and get whether the user is logged in.