Simplify login
There's no need to have a separate login page or automatically redirect from the home page to the login page. Just have the home page show the login button if the user is not logged in.

This way, you can still go back to whatever page you were previously if you happened to visit the home page and was redirected to login (there would be a loop where going back from the login page would redirect you back to the login page immediately since you're not signed in), and you still get the seamless login experience only having to press a single button.