Add rudimentary seeker; make pattern for passing get/setters
The seeker isn't perfect as it updates based upon `audio`'s `onTimeUpdate` event which only fires at most every 250ms. This results in a pretty choppy progress slider, but it works perfectly for now.

I've been annoyed with constantly needing to pass down getters and setters as props so I've established the pattern of just sending down the entire get/set state array returned from `useState` if a child requires it. This way there are fewer props to consume and it's just delegated to children to destructure them as needed.