Create `useAuthorizedFetch` hook
This is a slightly clunky pattern, but this hook will return a varation of `fetch` which handles unauthorized responses by setting the login state of the application to `LoggedOut`. API requests will need to take in this function as a parameter in order to properly handle unauthorized requests.

I'd like to figure out a better pattern that doesn't require this much work, but for now it works.