Check for OK response
The `fetch` API does not throw in the event of non-OK responses. In order for `react-query` to properly respond to errors and retry, an exception needs to be thrown.

In the future either I'll need to find a better request library that throws or make a good abstraction that ensures that they'll always throw. Right now, so long as I use `authorizedFetch` all should be good.