Add login state transition function
To fully model the intended login state machine of this app, it needs to be able to handle state transitions on top of state behavior. For instance, if a use gets logged out when they were originally logged in, a toast should be shown. Both the previous and new states are needed in order to prevent this toast appearing say, when the user's login is unknown and then changed to logged out.

To accomplish do this, `useLogin` needs to expose the ability to transition the login state. This way state transitions logic can still be centralized there.

In order to accomplish this, I also needed to generalize the usage of `useLogin` to simply expose the login state and any transition functions. It shouldn't for all usages fire a `me` request to check the login state when it's unknown.

This leads me to want to also remove the login check request from `useLogin` and just put it in the `App` component instead.