Try to guard against logout edge case
Currently, when the user gets logged out (possibly when on a page > 0), rendering fails because an out of bounds index error occurs on the sound data which likely gets wiped upon logout.

I've added a guard to make sure that if the user is not logged in, the access of the sound won't even occur.