Cleanup `TimeStamp` component; fix widths
The `TimeStamp` component has width issues when there are two minute digits. I've create a slightly better but still hacky and not fully tested way to get the right width for a time depending on how many characters it takes up, but this isn't perfect.

Ideally I should make a storybook for these components so that I can test them in different situations.