Stay ahead using sounds vs pages
Since page lengths are now variable (in case some sounds fail to be fetched), Raudi should stay ahead in terms of sounds rather than pages.

To avoid needing to constantly be counting the number of sounds since data is paginated, state is kept for a running total number of sounds along with the current sound index. This is simply updated when sounds come in, the user navigates to the next sound, or on logout so that whether or not the app is ahead by a decent amount of sounds can be calculated in constant time.