Hoist volume state out player
Taking a note out of the React beta docs: https://beta.reactjs.org/learn/you-might-not-need-an-effect#resetting-all-state-when-a-prop-changes

The player was originally meant to represent a single sound. This means that it should be more or less completely reset when the sound changes. This does not need to be done with an effect, since this would be at the cost of extra re-renders.

However, volume is part of the player component, but is intended to persist across sounds.

This leads me to the conclusion that the player doesn't necessarily represent a sound, but a component that plays a single sound. This starts to shed light on the difference between components and data. The player is a component. The sound is data. The player reacts to the sound changing, and has controlled aspects.

In order to reset the player, the sound's ID is given as its key. In addition, volume state is hoisted out of the component and now considered a prop so that it may persist through resets of the player when the sound changes.