Remove unnecessary effect from `useTruncation`
There's no need to reset the hook state when a prop changes using an effect. However, I'm not sure if this is the best pattern either, though it is semi-recommended by the React beta docs over using an effect (https://beta.reactjs.org/learn/you-might-not-need-an-effect). It's not clear what the ideal scenario is since a custom hook doesn't have a key, though maybe it doesn't matter since the hook is only used in terms of a component?

As long as that component "resets" when its state changes then the hook should theoretically do the same.